requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
fonttools==4.40.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langchain==0.0.186
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.8
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.7
wcwidth==0.2.6
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
